Fix bug where tail recursion makes forward jump miss
	 0; : t1 IF . THEN ;
	 0; : t2 ."t2" ;
	 0; 0 0- t1 drop ;
	t2 0;

The lines above show t1 jumping into the next function rather than
returning.  The tail recursion optimization mistakenly changes
the final call into a jump and omits the ret instruction.
Without that ret instruction, the forward jump of the IF lands
in the next function.  Here is the disassembly of t1:

   0x0804d7f4:  74 05   je     0x804d7fb
   0x0804d7f6:  e9 7d ef ff ff  jmp    0x804c778

This commit disables the optimization for a call immediately
preceding a THEN, REPEAT, UNTIL, or END.  The new disassembly
of t1 has a ret:

   0x0804d800:  74 05   je     0x804d807
   0x0804d802:  e8 7d ef ff ff  call   0x804c784
   0x0804d807:  c3      ret